requirements.txt
vllm>=0.5.0
torch>=2.1.0
transformers>=4.38.0
sentencepiece>=0.1.98
accelerate>=0.25.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.23.0
PyYAML>=6.0
python-dotenv>=1.0.0
tqdm>=4.64.0
psutil>=5.9.0
num2words